+++ core-ng/src/main/java/core/framework/api/module/QueueConfig.java
@@ -5,7 +5,7 @@ import core.framework.api.util.Types;
 import core.framework.impl.module.ModuleContext;
 import core.framework.impl.queue.MessageValidator;
 import core.framework.impl.queue.RabbitMQ;
-import core.framework.impl.queue.RabbitMQEndpoint;
+import core.framework.impl.queue.RabbitMQImpl;
 import core.framework.impl.queue.RabbitMQListener;
 import core.framework.impl.queue.RabbitMQPublisher;
 
@@ -16,48 +16,70 @@ import java.time.Duration;
  */
 public final class QueueConfig {
     private final ModuleContext context;
+    private final RabbitMQ rabbitMQ;
 
     public QueueConfig(ModuleContext context) {
         this.context = context;
-    }
-
-    public RabbitMQConfig rabbitMQ() {
-        if (context.queueManager.rabbitMQ == null) {
-            context.queueManager.rabbitMQ = new RabbitMQ();
-            if (!context.isTest()) {
-                context.backgroundTask().scheduleWithFixedDelay(context.queueManager.rabbitMQ.pool::refresh, Duration.ofMinutes(5));
-                context.shutdownHook.add(context.queueManager.rabbitMQ::close);
+        if (context.beanFactory.registered(RabbitMQ.class, null)) {
+            rabbitMQ = context.beanFactory.bean(RabbitMQ.class, null);
+        } else {
+            if (context.isTest()) {
+                rabbitMQ = context.mockFactory.create(RabbitMQ.class);
+            } else {
+                RabbitMQImpl rabbitMQ = new RabbitMQImpl();
+                context.backgroundTask().scheduleWithFixedDelay(rabbitMQ.pool::refresh, Duration.ofMinutes(5));
+                context.shutdownHook.add(rabbitMQ::close);
+                this.rabbitMQ = rabbitMQ;
             }
+            context.beanFactory.bind(RabbitMQ.class, null, rabbitMQ);
         }
-        return new RabbitMQConfig(context);
     }
 
-    public MessageHandlerConfig subscribe(String queueURI) {
-        return context.queueManager.listeners().computeIfAbsent(queueURI, this::listener);
+    public MessageHandlerConfig subscribe(String queue) {
+        return context.queueManager.listeners().computeIfAbsent(queue, key -> {
+            RabbitMQListener listener = new RabbitMQListener(rabbitMQ, queue, context.executor, context.queueManager.validator(), context.logManager);
+            if (!context.isTest()) {
+                context.startupHook.add(listener::start);
+                context.shutdownHook.add(listener::stop);
+            }
+            return listener;
+        });
     }
 
-    public <T> void publish(String destination, Class<T> messageClass) {
+    public <T> void publish(String exchange, String routingKey, Class<T> messageClass) {
         MessageValidator validator = context.queueManager.validator();
         validator.register(messageClass);
-        MessagePublisher<T> publisher = publisher(destination, messageClass);
+        MessagePublisher<T> publisher = new RabbitMQPublisher<>(rabbitMQ, exchange, routingKey, messageClass, context.queueManager.validator(), context.logManager);
         context.beanFactory.bind(Types.generic(MessagePublisher.class, messageClass), null, publisher);
     }
 
-    private MessageHandlerConfig listener(String queueURI) {
-        RabbitMQListener listener = new RabbitMQListener(context.queueManager.rabbitMQ(), new RabbitMQEndpoint(queueURI).routingKey, context.executor, context.queueManager.validator(), context.logManager);
+    public void hosts(String... hosts) {
         if (!context.isTest()) {
-            context.startupHook.add(listener::start);
-            context.shutdownHook.add(listener::stop);
+            ((RabbitMQImpl) rabbitMQ).hosts(hosts);
         }
-        return listener;
     }
 
-    @SuppressWarnings("unchecked")
-    private <T> MessagePublisher<T> publisher(String uri, Class<T> messageClass) {
-        if (context.isTest()) {
-            return context.mockFactory.create(MessagePublisher.class, uri, context.queueManager.validator());
-        } else {
-            return new RabbitMQPublisher<>(context.queueManager.rabbitMQ(), new RabbitMQEndpoint(uri), messageClass, context.queueManager.validator(), context.logManager);
+    public void user(String user) {
+        if (!context.isTest()) {
+            ((RabbitMQImpl) rabbitMQ).user(user);
+        }
+    }
+
+    public void password(String password) {
+        if (!context.isTest()) {
+            ((RabbitMQImpl) rabbitMQ).password(password);
+        }
+    }
+
+    public void timeout(Duration timeout) {
+        if (!context.isTest()) {
+            ((RabbitMQImpl) rabbitMQ).timeout(timeout);
+        }
+    }
+
+    public void poolSize(int minSize, int maxSize) {
+        if (!context.isTest()) {
+            ((RabbitMQImpl) rabbitMQ).pool.size(minSize, maxSize);
         }
     }
 }