+++ core-ng/src/main/java/core/framework/api/queue/MessagePublisher.java
@@ -6,6 +6,5 @@ package core.framework.api.queue;
 public interface MessagePublisher<T> {
     void publish(T message);
 
-    // RabbitMQ uses routingKey as queueName with default exchange, SNS/SQS don't support routingKey
-    void publish(String routingKey, T message);
+    void publish(String exchange, String routingKey, T message);
 }