+++ core-ng/src/main/java/core/framework/impl/module/QueueManager.java
@@ -3,7 +3,6 @@ package core.framework.impl.module;
 import core.framework.api.module.MessageHandlerConfig;
 import core.framework.api.util.Maps;
 import core.framework.impl.queue.MessageValidator;
-import core.framework.impl.queue.RabbitMQ;
 
 import java.util.Map;
 
@@ -11,7 +10,6 @@ import java.util.Map;
  * @author neo
  */
 public class QueueManager {
-    public RabbitMQ rabbitMQ;
     private MessageValidator validator;
     private Map<String, MessageHandlerConfig> listeners;
 
@@ -27,12 +25,5 @@ public class QueueManager {
             validator = new MessageValidator();
         }
         return validator;
-    }
-
-    public RabbitMQ rabbitMQ() {
-        if (rabbitMQ == null) {
-            throw new Error("rabbitMQ is not configured, please use queue().rabbitMQ() to configure");
-        }
-        return rabbitMQ;
     }
 }