+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Change log
 
+### 4.1.3 (2/11/2016)
+* db: lower default db timeout to 15s
+* queue: updated rabbitMQ api and config api
+
 ### 4.1.2 (2/10/2016)
 * web: removed web/not-found, web/method-not-allowed action assign, since we use error_code now
 * background: moved pool-cleanup, collect-stat job to background thread, not included in action